+++ patches/Accessibility/strings_it.cpp
@@ -484,6 +484,10 @@ const char* Get(Id id) {
         case Id::FmtModSettingSlider:         return "%s: %d percento";
         case Id::ModSettingUrgentVolume:      return "Volume degli annunci vocali";
         case Id::ModSettingUrgentVolumePreview: return "Annuncio di esempio";
+        case Id::ModSettingSupportModder:           return "Sostieni il modder";
+        case Id::ModSettingLatestChanges:           return "Ultime modifiche";
+        case Id::ModSettingLinkOpened:              return "Apertura nel browser.";
+        case Id::ModSettingLinkFailed:              return "Impossibile aprire il link.";
 
         case Id::ModSettingAudioGlossary:           return "Glossario audio";
         case Id::ModSettingsAudioGlossaryOpened:    return "Glossario audio aperto";